feat: calculate_metrics in Endpoint.jsx
