+++ client/src/components/Editor.jsx
@@ -9,6 +9,7 @@ import "../styles/Editor.css";
 import { gql } from "graphql-tag";
 import Split from "react-split";
 import { DEFAULT_EDITOR_OPTIONS } from "../utils/defaultEditorOptions";
+import { calculate_metrics } from "../utils/metrics";
 
 /* Default Initial Display for Query Operations */
 const defaultOperations =
@@ -226,79 +227,6 @@ export default function Editor({ schema, endpoint, setQuery }) {
     });
   }
 
-  /*
-    metrics feedback
-  */
-  function calculate_metrics() {
-    // Check performance support
-    if (performance === undefined) {
-      console.log("= Calculate Load Times: performance NOT supported");
-      return;
-    }
-
-    // Get a list of "resource" performance entries
-    const resources = performance.getEntriesByName(endpoint);
-    if (resources === undefined || resources.length <= 0) {
-      console.log(
-        "= Calculate Load Times: there are NO `resource` performance records"
-      );
-      return;
-    }
-    // console.log("all resources: ", resources);
-    const mostRecentQuery = resources.pop();
-    console.log(mostRecentQuery);
-    let t =
-      mostRecentQuery.fetchStart > 0
-        ? mostRecentQuery.responseEnd - mostRecentQuery.fetchStart
-        : "0";
-    console.log(
-      `Total query time for ${mostRecentQuery.name} - ${mostRecentQuery.initiatorType} (including queuing) = ${t}`
-    );
-
-    // // filter resources to only contain entries of those that reference the endpoint
-    // const endpointResources = resources.filter((resource) => {
-    //   return resource.name === endpoint;
-    // });
-    // console.log("endpointResources:", endpointResources);
-
-    // console.log("= Calculate Load Times");
-    // resources.forEach((resource, i) => {
-    //   console.log(
-    //     `== Resource[${i}] - ${resource.name} - ${resource.initiatorType}`
-    //   );
-    //   // Fetch until response end
-    //   let t =
-    //     resource.fetchStart > 0
-    //       ? resource.responseEnd - resource.fetchStart
-    //       : "0";
-    //   console.log(`… Total query time (including queuing) = ${t}`);
-    // });
-
-    // // SIZE
-    // // For each "resource", display its *Size property values
-    // console.log("= Display Size Data");
-    // endpointResources.forEach((entry, i) => {
-    //   console.log(`== Resource[${i}] - ${entry.name}`);
-    //   if ("decodedBodySize" in entry) {
-    //     console.log(`… decodedBodySize[${i}] = ${entry.decodedBodySize}`);
-    //   } else {
-    //     console.log(`… decodedBodySize[${i}] = NOT supported`);
-    //   }
-
-    //   if ("encodedBodySize" in entry) {
-    //     console.log(`… encodedBodySize[${i}] = ${entry.encodedBodySize}`);
-    //   } else {
-    //     console.log(`… encodedBodySize[${i}] = NOT supported`);
-    //   }
-
-    //   if ("transferSize" in entry) {
-    //     console.log(`… transferSize[${i}] = ${entry.transferSize}`);
-    //   } else {
-    //     console.log(`… transferSize[${i}] = NOT supported`);
-    //   }
-    // });
-  }
-
   /* Get Operations & Validate
      Return: {valid:Boolean <, error:String, operationString:String, operationType:String>} */
   const getOperationsAndValidate = () => {
@@ -366,7 +294,7 @@ export default function Editor({ schema, endpoint, setQuery }) {
     // Note: this app only supports a single iteration for http GET/POST,
     // no multipart or subscriptions yet.
     const data = await result.next();
-    calculate_metrics();
+    calculate_metrics(endpoint);
 
     // Display the results in results pane
     resultsModel?.setValue(
